requirements.txt
# Core dependencies
numpy
opencv-python
Pillow

# Scientific computing
scikit-learn
scipy

# Visualization
matplotlib
seaborn

# Performance accelerators
numba
llvmlite

# Development/testing
pytest
pytest-cov

# Optional (but recommended)
imageio
tqdm